requirements.txt
requests
openai
dotenv
schedule
time
csv
snowflake-connector-python